norm

+++ ft_strnlen.c
@@ -6,14 +6,14 @@
 /*   By: epoggio <marvin@le-101.fr>                 +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/10/11 18:24:18 by epoggio      #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/11 18:24:18 by epoggio     ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/10/11 19:46:23 by epoggio     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_strnlen(const char *str, size_t n)
+size_t	ft_strnlen(const char *str, size_t n)
 {
 	int i;
